Add test to create a new user
* Added a new test scenario to create a dummy user first and then
calling the "get user by ID" controller to assert it has been saved
correctly.
* Added the API access token to the `test.properties` files. This is not
how it usually should be done. Access tokens and other secret
credentials should be stored securely in the vault on the build server
and should not be checked in to the VCS. I've committed my API access
token for demo purposes only.