Add tests for Android
Added three scenarios for HSL Android application. The
`whenUserOpensLogInPage_thenUserCanSeeForgotYourPasswordLink` scenario
is failing on purpose. That is done as an example to show the failed
test in report.
In this example I follow the Page Object pattern to organize all page
elements and components on respective pages/screens of the app. That
makes the code look cleaner in the test and reduces duplication.

Also, migrate from Maven to Gradle since Gradle works better with
Android Studio.